pyproject.toml
[project]
name = "nonebot-plugin-llnonebot-api"
version = "0.1.0"
description = "为 LLNoneBot 提供 API 服务 / Provide API services for LLNoneBot"
authors = [
    { name = "KomoriDev", email = "mute231010@gmail.com" },
]
dependencies = [
    "nonebot2[fastapi]>=2.3.3",
    "fastapi>=0.115.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.nonebot]
plugins = ["nonebot_plugin_llnonebot_api"]

[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.6.1",
    "black>=24.4.2",
    "isort>=5.13.2",
    "pre-commit>=3.7.1",
]

[tool.black]
line-length = 90
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 90
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.ruff]
line-length = 90
target-version = "py39"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901"]

[tool.pyright]
pythonVersion = "3.9"
pythonPlatform = "All"
typeCheckingMode = "basic"

